requirements.txt
openai>=1.30.0
pandas>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.4
tqdm>=4.66.0
